+++ ruff_cli/src/iterators.rs
