+++ ruff_cli/src/lib.rs
@@ -1,0 +1,6 @@
+#![allow(clippy::must_use_candidate, dead_code)]
+
+mod cli;
+
+// used by ruff_dev::generate_cli_help
+pub use cli::Cli;
