+++ ruff_cli/src/main.rs
@@ -1,11 +1,17 @@
+#![allow(
+    clippy::match_same_arms,
+    clippy::missing_errors_doc,
+    clippy::module_name_repetitions,
+    clippy::too_many_lines
+)]
+#![forbid(unsafe_code)]
+
 use std::io::{self};
 use std::path::{Path, PathBuf};
 use std::process::ExitCode;
 use std::sync::mpsc::channel;
 
-use ::ruff::cli::{extract_log_level, Cli, Overrides};
 use ::ruff::logging::{set_up_logging, LogLevel};
-use ::ruff::printer::{Printer, Violations};
 use ::ruff::resolver::{
     resolve_settings_with_processor, ConfigProcessor, FileDiscovery, PyprojectDiscovery, Relativity,
 };
@@ -13,13 +19,23 @@ use ::ruff::settings::configuration::Configuration;
 use ::ruff::settings::types::SerializationFormat;
 use ::ruff::settings::{pyproject, Settings};
 #[cfg(feature = "update-informer")]
-use ::ruff::updates;
-use ::ruff::{commands, fix, fs, warn_user_once};
+use ::ruff::{fix, fs, warn_user_once};
 use anyhow::Result;
 use clap::{CommandFactory, Parser};
+use cli::{extract_log_level, Cli, Overrides};
 use colored::Colorize;
 use notify::{recommended_watcher, RecursiveMode, Watcher};
 use path_absolutize::path_dedot;
+use printer::{Printer, Violations};
+
+mod cache;
+mod cli;
+mod commands;
+mod diagnostics;
+mod iterators;
+mod printer;
+#[cfg(all(feature = "update-informer"))]
+pub mod updates;
 
 /// Resolve the relevant settings strategy and defaults for the current
 /// invocation.
@@ -72,7 +88,7 @@ fn resolve(
     }
 }
 
-pub(crate) fn inner_main() -> Result<ExitCode> {
+pub fn main() -> Result<ExitCode> {
     // Extract command-line arguments.
     let (cli, overrides) = Cli::parse().partition();
     let log_level = extract_log_level(&cli);
@@ -130,7 +146,7 @@ pub(crate) fn inner_main() -> Result<ExitCode> {
     };
 
     if let Some(code) = cli.explain {
-        commands::explain(&code, &format)?;
+        commands::explain(&code, format)?;
         return Ok(ExitCode::SUCCESS);
     }
     if cli.show_settings {
@@ -196,7 +212,7 @@ pub(crate) fn inner_main() -> Result<ExitCode> {
             cache.into(),
             fix::FixMode::None,
         )?;
-        printer.write_continuously(&messages)?;
+        printer.write_continuously(&messages);
 
         // Configure the file watcher.
         let (tx, rx) = channel();
@@ -226,7 +242,7 @@ pub(crate) fn inner_main() -> Result<ExitCode> {
                             cache.into(),
                             fix::FixMode::None,
                         )?;
-                        printer.write_continuously(&messages)?;
+                        printer.write_continuously(&messages);
                     }
                 }
                 Err(err) => return Err(err.into()),
